solution/adspresso/requirements.txt
adrf==0.1.9
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-property==0.2.2
attrs==25.1.0
billiard==4.2.1
blinker==1.9.0
botocore==1.36.24
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
confluent-kafka==2.8.0
coverage==7.6.12
cron-descriptor==1.4.5
DAWG-Python==0.7.2
Django==5.1.6
django-celery-beat==2.7.0
django-filter==24.3
django-minio-backend==3.8.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
docopt==0.6.2
drf-yasg==1.21.8
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
frozenlist==1.5.0
fsspec==2025.2.0
gevent==24.11.1
geventhttpclient==2.3.3
gigachat==0.1.38
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
intervaltree==3.1.0
ipymarkup==0.9.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
locust==2.32.9
Markdown==3.7
MarkupSafe==3.0.2
minio==7.2.15
msgpack==1.1.0
multidict==6.1.0
natasha==1.6.0
navec==0.10.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pytest==8.3.4
pytest-django==4.10.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
razdel==0.5.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.2
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
slovnet==0.6.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yargy==0.16.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
